// packages/medusa-fulfillment-apiship/src/apiship-client/models/CancelOrderResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CancelOrderResponse = {
    /**
     * Номер заказа
     */
    orderId?: number;
    /**
     * Дата отмены заказа
     */
    canceled?: string;
};