// packages/medusa-fulfillment-apiship/src/apiship-client/models/DeleteOrderResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DeleteOrderResponse = {
    /**
     * Номер заказа
     */
    orderId?: number;
    /**
     * Дата удаления заказа
     */
    deleted?: string;
};