// packages/medusa-fulfillment-apiship/src/apiship-client/models/OrderResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type OrderResponse = {
    /**
     * Номер заказа
     */
    orderId?: number;
    /**
     * Дата создания заказа
     */
    created?: string;
};