// packages/medusa-fulfillment-apiship/src/apiship-client/models/OrderSyncResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type OrderSyncResponse = {
    /**
     * Номер заказа
     */
    orderId?: number;
    /**
     * Номер заказа в системе службы доставки
     */
    providerNumber?: string;
    /**
     * Дополнительный номер заказа в системе службы доставки
     */
    additionalProviderNumber?: string;
    /**
     * Дата создания заказа
     */
    created?: string;
};